Consolidate to a single NoOpTelemetryReporter
